# swagger_server/swagger/swagger.yaml
...
swagger: "2.0"
info:
  version: "0.0.2"
  title: "Optimization framework service"
  license:
    name: "Apache"
host: "localhost:8080"
basePath: "/v1"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /inputs/mqtt:
    get:
      tags:
      - "inputs"
      summary: "Receives all mqtt data from the framework"
      description: ""
      operationId: "get_all_mqtt_data_source_values"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/MQTTOutputAll"
        400:
          description: "Bad request"
        405:
          description: "Invalid input"
        500:
          description: "Internal server error"
      x-swagger-router-controller: "swagger_server.controllers.inputs_controller"
    post:
      tags:
      - "inputs"
      summary: "Creates a new mqtt data source as input"
      description: ""
      operationId: "mqtt_input_source"
      consumes:
      - "application/json"
      - "text/plain"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "MQTT_Input_Source"
        description: "Data source to be created"
        required: true
        schema:
          allOf:
          - $ref: "#/definitions/MQTTInputSource"
          - $ref: "#/definitions/Generic_Files"
      responses:
        201:
          description: "Created succesfully"
          headers:
            Location:
              type: "string"
              description: "Id referencing the created instance."
        400:
          description: "Bad request"
        405:
          description: "Invalid input"
        500:
          description: "Internal server error"
      x-swagger-router-controller: "swagger_server.controllers.inputs_controller"
  /inputs/mqtt/ids:
    get:
      tags:
      - "inputs"
      summary: "Receives data from the framework"
      description: ""
      operationId: "get_all_mqtt_data_source_ids"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/OutputIdsList"
        400:
          description: "Bad request"
        405:
          description: "Invalid input"
        500:
          description: "Internal server error"
      x-swagger-router-controller: "swagger_server.controllers.inputs_controller"
  /inputs/mqtt/{id}:
    get:
      tags:
      - "inputs"
      summary: "Receives data from the framework"
      description: ""
      operationId: "get_mqtt_data_source_values"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of the data source"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/MQTTInputSource"
        400:
          description: "Bad request"
        405:
          description: "Invalid input"
        500:
          description: "Internal server error"
      x-swagger-router-controller: "swagger_server.controllers.inputs_controller"
    put:
      tags:
      - "inputs"
      summary: "Submits data to the framework"
      description: ""
      operationId: "mqtt_input_put"
      consumes:
      - "application/json"
      - "text/plain"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of the data source"
        required: true
        type: "string"
      - in: "body"
        name: "dataset"
        description: "Dataset submitted to the framework"
        required: true
        schema:
          allOf:
          - $ref: "#/definitions/MQTTInputSource"
          - $ref: "#/definitions/Generic_Files"
      responses:
        200:
          description: "Created succesfully"
        400:
          description: "Bad request"
        405:
          description: "Invalid input"
        500:
          description: "Internal server error"
      x-swagger-router-controller: "swagger_server.controllers.inputs_controller"
    delete:
      tags:
      - "inputs"
      summary: "Deletes the loaded data"
      description: ""
      operationId: "delete_mqtt_registry"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Name of the registry to be deleted"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
        400:
          description: "Bad request"
        405:
          description: "Invalid input"
        500:
          description: "Internal server error"
      x-swagger-router-controller: "swagger_server.controllers.inputs_controller"
  /inputs/dataset:
    get:
      tags:
      - "inputs"
      summary: "Receives data from the framework"
      description: ""
      operationId: "get_all_data_source_values"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/FileOutputAll"
        400:
          description: "Bad request"
        405:
          description: "Invalid input"
        500:
          description: "Internal server error"
      x-swagger-router-controller: "swagger_server.controllers.inputs_controller"
    post:
      tags:
      - "inputs"
      summary: "Creates a new data source as input"
      description: ""
      operationId: "dataset_input_source"
      consumes:
      - "application/json"
      - "text/plain"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "File_Input_Source"
        description: "Dataset submitted to the framework"
        required: true
        schema:
          $ref: "#/definitions/FileInputSource"
      responses:
        201:
          description: "Created succesfully"
          headers:
            Location:
              type: "string"
              description: "Id referencing the created instance."
        400:
          description: "Bad request"
        405:
          description: "Invalid input"
        500:
          description: "Internal server error"
      x-swagger-router-controller: "swagger_server.controllers.inputs_controller"
  /inputs/dataset/ids:
    get:
      tags:
      - "inputs"
      summary: "Receives data from the framework"
      description: ""
      operationId: "get_all_data_source_ids"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/OutputIdsList"
        400:
          description: "Bad request"
        405:
          description: "Invalid input"
        500:
          description: "Internal server error"
      x-swagger-router-controller: "swagger_server.controllers.inputs_controller"
  /inputs/dataset/{id}:
    get:
      tags:
      - "inputs"
      summary: "Receives data from the framework"
      description: ""
      operationId: "get_data_source_values"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of the data source"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/FileInputSource"
        400:
          description: "Bad request"
        405:
          description: "Invalid input"
        500:
          description: "Internal server error"
      x-swagger-router-controller: "swagger_server.controllers.inputs_controller"
    put:
      tags:
      - "inputs"
      summary: "Submits data to the framework"
      description: ""
      operationId: "dataset_input_put"
      consumes:
      - "application/json"
      - "text/plain"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of the data source"
        required: true
        type: "string"
      - in: "body"
        name: "dataset"
        description: "Dataset submitted to the framework"
        required: true
        schema:
          $ref: "#/definitions/FileInputSource"
      responses:
        200:
          description: "OK"
        400:
          description: "Bad request"
        405:
          description: "Invalid input"
        500:
          description: "Internal server error"
      x-swagger-router-controller: "swagger_server.controllers.inputs_controller"
    delete:
      tags:
      - "inputs"
      summary: "Deletes the loaded data"
      description: ""
      operationId: "delete_dataset_registry"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Name of the registry to be deleted"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
        400:
          description: "Bad request"
        405:
          description: "Invalid input"
        500:
          description: "Internal server error"
      x-swagger-router-controller: "swagger_server.controllers.inputs_controller"
  /inputs/{id}:
    delete:
      tags:
      - "inputs"
      summary: "Deletes loaded data by id"
      description: ""
      operationId: "delete_data_source_all"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Id of the data source to be deleted"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
        400:
          description: "Bad request"
        405:
          description: "Invalid input"
        500:
          description: "Internal server error"
      x-swagger-router-controller: "swagger_server.controllers.inputs_controller"
...
definitions:
...
  Generic_Files:
    type: "object"
    example:
      custom_name_1:
      - 1200
      - 12300
      custom_name_2:
      - 1500
      - 1300
    additionalProperties:
      type: "array"
      items:
        type: "number"
...
  FileOutputAll:
    type: "object"
    example:
      12asd123d:
        generic:
          custom_name_1:
          - 1200
          - 12300
          custom_name_2:
          - 1500
          - 1300
        load:
          P_Load_R:
          - 705.3999999999999772626324556767940521240234375
          - 305.6000000000000227373675443232059478759765625
          Q_Load_T:
          - 705.3999999999999772626324556767940521240234375
          - 305.6000000000000227373675443232059478759765625
          P_Load_S:
          - 705.3999999999999772626324556767940521240234375
          - 305.6000000000000227373675443232059478759765625
          Q_Load_S:
          - 705.3999999999999772626324556767940521240234375
          - 305.6000000000000227373675443232059478759765625
          P_Load_T:
          - 705.3999999999999772626324556767940521240234375
          - 305.6000000000000227373675443232059478759765625
          Q_Load:
          - 705.3999999999999772626324556767940521240234375
          - 305.6000000000000227373675443232059478759765625
          P_Load:
          - 705.3999999999999772626324556767940521240234375
          - 305.6000000000000227373675443232059478759765625
          Q_Load_R:
          - 705.3999999999999772626324556767940521240234375
          - 305.6000000000000227373675443232059478759765625
        ESS:
          SoC_Value: 41
          meta:
            ESS_Max_Discharge_Power: 2500
            ESS_Max_SoC: 90
            ESS_Charging_Eff: 0.90000000000000002220446049250313080847263336181640625
            ESS_Min_SoC: 20
            ESS_Max_Charge_Power: 2500
            ESS_Discharging_Eff: 0.84999999999999997779553950749686919152736663818359375
            ESS_Capacity: 3560
        grid:
          meta:
            Q_Grid_Max_Export_Power: 750
            Min_Voltage_Drop: 5
            P_Grid_Max_Export_Power: 1550
            Max_Voltage_Drop: 10
        photovoltaic:
          P_PV:
          - 705.3999999999999772626324556767940521240234375
          - 305.6000000000000227373675443232059478759765625
        global_control:
          ESS_Control:
          - 1300.200000000000045474735088646411895751953125
          - 1200.5
      45dsf785s:
        load:
          Q_Load:
          - 705.3999999999999772626324556767940521240234375
          - 305.6000000000000227373675443232059478759765625
          P_Load:
          - 705.3999999999999772626324556767940521240234375
          - 305.6000000000000227373675443232059478759765625
        ESS:
          SoC_Value: 41
          meta:
            ESS_Max_Discharge_Power: 2500
            ESS_Max_SoC: 90
            ESS_Charging_Eff: 0.90000000000000002220446049250313080847263336181640625
            ESS_Min_SoC: 20
            ESS_Max_Charge_Power: 2500
            ESS_Discharging_Eff: 0.84999999999999997779553950749686919152736663818359375
            ESS_Capacity: 3560
        grid:
          meta:
            Q_Grid_Max_Export_Power: 750
            Min_Voltage_Drop: 5
            P_Grid_Max_Export_Power: 1550
            Max_Voltage_Drop: 10
        photovoltaic:
          P_PV:
          - 705.3999999999999772626324556767940521240234375
          - 305.6000000000000227373675443232059478759765625
        global_control:
          ESS_Control:
          - 1300.200000000000045474735088646411895751953125
          - 1200.5
    additionalProperties:
      $ref: "#/definitions/FileInputSource"
  OutputIdsList:
    type: "array"
    items:
      type: "string"
    example:
    - "a123asd456"
    - "b456fgh789"
    - "c852edc963"
...